Added functionality to update multiple employee's salary in the database using threads and thread logs shown on the console

+++ src/main/java/com/capgemini/payrolldata/EmployeePayrollData.java
@@ -51,7 +51,7 @@ public class EmployeePayrollData {
 		this.empName = empName;
 	}
 
-	public long getEmpId() {
+	public int getEmpId() {
 		return empId;
 	}
